[update] - atualização no nome do pacote para publicação na PlayStore.
