+++ app/src/androidTest/java/br/com/phmiranda/agenda/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package br.com.alura.agenda;
+package br.com.phmiranda.agenda;
 
 import android.content.Context;
