+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="br.com.alura.agenda">
+    package="br.com.phmiranda.agenda">
 
     <application
         android:allowBackup="true"
@@ -9,12 +9,12 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.ListaAlunosActivity">
+        <activity android:name="br.com.phmiranda.agenda.view.ListaAlunosActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.FormularioAlunoActivity" />
+        <activity android:name="br.com.phmiranda.agenda.view.FormularioAlunoActivity" />
     </application>
 </manifest>
