--- app/src/main/java/br/com/alura/agenda/api/AlunoResource.java
@@ -1,4 +1,0 @@
-package br.com.alura.agenda.api;
-
-public class AlunoResource {
-}
