--- app/src/main/java/br/com/alura/agenda/config/AppBuilder.java
@@ -1,4 +1,0 @@
-package br.com.alura.agenda.config;
-
-public class AppBuilder {
-}
