--- app/src/main/java/br/com/alura/agenda/repository/AlunoRepository.java
@@ -1,9 +1,0 @@
-package br.com.alura.agenda.repository;
-
-import br.com.alura.agenda.entity.Aluno;
-
-public class AlunoRepository {
-    public void salvar(Aluno aluno){
-
-    }
-}
