--- app/src/main/java/br/com/alura/agenda/services/AlunoService.java
@@ -1,4 +1,0 @@
-package br.com.alura.agenda.services;
-
-public class AlunoService {
-}
