+++ app/src/main/java/br/com/phmiranda/agenda/api/AlunoResource.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.api;
+
+public class AlunoResource {
+}
