+++ app/src/main/java/br/com/phmiranda/agenda/config/AppBuilder.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.config;
+
+public class AppBuilder {
+}
