+++ app/src/main/java/br/com/phmiranda/agenda/repository/AlunoRepository.java
@@ -1,0 +1,9 @@
+package br.com.phmiranda.agenda.repository;
+
+import br.com.phmiranda.agenda.entity.Aluno;
+
+public class AlunoRepository {
+    public void salvar(Aluno aluno){
+
+    }
+}
