+++ app/src/main/java/br/com/phmiranda/agenda/services/AlunoService.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.services;
+
+public class AlunoService {
+}
