+++ app/src/main/java/br/com/phmiranda/agenda/view/FormularioAlunoActivity.java
@@ -1,18 +1,16 @@
-package br.com.alura.agenda.view;
+package br.com.phmiranda.agenda.view;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
-import br.com.alura.agenda.R;
-import br.com.alura.agenda.entity.Aluno;
-import br.com.alura.agenda.repository.AlunoDao;
+import br.com.phmiranda.agenda.R;
+import br.com.phmiranda.agenda.entity.Aluno;
+import br.com.phmiranda.agenda.repository.AlunoDao;
 
 public class FormularioAlunoActivity extends AppCompatActivity {
     public static final String FORMULARIO_TITULO = "Formulário de Aluno";
