+++ app/src/test/java/br/com/phmiranda/agenda/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package br.com.alura.agenda;
+package br.com.phmiranda.agenda;
 
 import org.junit.Test;
